requirements.txt
sympy>=1.0
python-afl>=0.7.3
pytest>=8.0